use size_t

+++ rdma_client/data_format.h
@@ -1,6 +1,8 @@
+#include <cstdint>
+
 char table_name[16];
 
 struct size_pair {
-  int data_size;      // size of final data (table)
-  int metadata_size;  // size of metadata
+  size_t data_size;      // size of final data (table)
+  size_t metadata_size;  // size of metadata
 } __attribute__ ((packed));
